+++ test/dummy/app/models/workspace.rb
@@ -2,4 +2,5 @@ class Workspace < ApplicationRecord
   recording_studio_recordable label: "Workspace", root: true
   include RecordingStudio::Capabilities::Example.to(label: "dummy workspace")
   RecordingStudio.enable_capability(:accessible, on: self) if defined?(RecordingStudioAccessible)
+  RecordingStudio.enable_capability(:api_access_point, on: self) if defined?(RecordingStudioApi)
 end